requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
ldap3==2.9.1
jinja2==3.1.3
python-multipart==0.0.9
itsdangerous==2.1.2  # Required for SessionMiddleware
python-jose==3.3.0   # For JWT tokens if needed
